moved types and getUnixTime to time util
