+++ src/utils/time/index.ts
@@ -1,0 +1,11 @@
+export function getUNIXTime(dateTime: Date) {
+    if (!dateTime || isNaN(dateTime.getTime())) {
+        return null;
+    }
+    const unixTime = dateTime.getTime() * 1000;
+    if (unixTime <= 0) {
+        return null;
+    } else {
+        return unixTime;
+    }
+}
